Workaround for building with Anaconda's gfortran on El Capitan
Anaconda's gfortran (distributed withing the `gcc' conda package) comes with
runtime libraries (namely, libgfortran) that have install_names defined as
'@rpath/<library_name>'. These names get copied into the binaries that the
compiler builds. For the binaries to run, they need to know where @rpath
(defined as Anaconda's lib directory) is. When a conda package is built
using conda-build, an afterburner is run to add the necessary rpaths to all
built binaries. But before that happens, we'd typically set
DYLD_LIBRARY_PATH (or its FALLBACK variant) to let the binaries find the
necesary libraries.

El Capitan's SIP made this scheme difficult to maintain as *DYLD* variables
are clensed from the environment when anything in /usr/bin is run (e.g.,
make).  We therefore cannot just set DYLD_LIBRARY_PATH and have the build
succeed and get to the afterburner stage.

Instead, we detect and add the RPATH manually to the built libraries and
binaries. The RPATH is detected by inspecting the location of the gfortran
compiler, and guessing where its runtime library is (the logic is in
get-rpath.sh). This is then added to gfortran compiler/linker options in the
Makefiles.

This makes it possible to build oorb with Anaconda's gfortran, both with
EUPS and when building it using conda build.

n.b: I believe the correct fix for this issue is for Anaconda's gfortran
runtime libraries to have their install_name's be absolute (not
@rpath-relative) paths. I'll open an issue upstream.